+++ spec/views/my/page.html.erb_spec.rb
@@ -32,8 +32,7 @@ describe 'my/page' do
 
     render
 
-    assert_select 'tr.time-entry td.subject' do |td|
-      td.should have_link("#{issue.type.name} ##{issue.id}", :href => work_package_path(issue))
-    end
+    response.should have_selector("tr.time-entry td.subject a[href='#{work_package_path(issue)}']",
+                                  :text => "#{issue.type.name} ##{issue.id}")
   end
 end