+++ doc/CHANGELOG.md
@@ -1,7 +1,21 @@
+<!---- copyright
+OpenProject is a project management system.
+
+Copyright (C) 2012-2013 the OpenProject Team
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License version 3.
+
+See doc/COPYRIGHT.rdoc for more details.
+
+++-->
+
 # Changelog
 
 * `#1418` Additional changes: Change links to issues/planning elements to use work_packages controller
 * `#1898` Separate action for changing wiki parent page (was same as rename before)
+* `#1923` Add permission that allows hiding repository statistics on commits per author
+* `#1850` Disable atom feeds via setting
 
 ## 3.0.0pre15
 
@@ -113,3 +127,4 @@
 * `#778` Integrate ban of former passwords
 * `#1209` Fix adding watcher to issue
 * `#1034` Create changelog and document format
+